system dockerized

+++ backend/api/views/user.py
@@ -196,6 +196,7 @@ class UserProfileCreatorViewSet(viewsets.ModelViewSet):
 
         user_activation_link = UserActivationLink(user=user)
         user_activation_link.save()
+
         EmailSender.send_user_activation_email(
             user, FrontendLinks.get_user_activation_link(user_activation_link)
         )
@@ -254,11 +255,11 @@ class LoginViewSet(viewsets.ViewSet):
             # decrypt keys with users password (or: if not encrypted atm, encrypt them with users password)
             private_key = encryption_keys.decrypt_private_key(user_password)
 
-            from backend.recordmanagement.helpers import (
-                resolve_missing_record_key_entries,
-            )
+            # from backend.recordmanagement.helpers import (
+            #     resolve_missing_record_key_entries,
+            # )
 
-            resolve_missing_record_key_entries(token.user, private_key)
+            # resolve_missing_record_key_entries(token.user, private_key)
             # TODO: superuser?
             if not token.user.is_superuser:
                 from backend.api.helpers import resolve_missing_rlc_keys_entries
@@ -270,6 +271,11 @@ class LoginViewSet(viewsets.ViewSet):
 
     def get(self, request):
         token = request.META["HTTP_AUTHORIZATION"].split(" ")[1]
+        try:
+            Token.objects.get(key=token)
+        except:
+            raise CustomError(ERROR__API__NOT_AUTHENTICATED)
+
         return Response(LoginViewSet.get_login_data(token))
 
     @staticmethod
@@ -327,7 +333,12 @@ class LoginViewSet(viewsets.ViewSet):
 
 
 class LogoutViewSet(APIView):
+    authentication_classes = ()
+    permission_classes = ()
+
     def post(self, request):
+        if not request.user.is_authenticated:
+            return Response()
         Token.objects.filter(user=request.user).delete()
         return Response()
